GetProgrammText & freeProgrammText

+++ PIC/main.cpp
@@ -10,9 +10,9 @@ int main(int argc, char *argv[]) {
 	b->setCommandSpeed(1000);
 	b->LoadProgramm("..\\Debug\\Testprogramme\\TPicSim8.LST");
 	//prog 7
-	//printf("Set Breakpoint returns: %04d, %04d\n",b->ToggleBreakpoint(40)+1,b->ToggleBreakpoint(55)+1);
+	//printf("Set Breakpoint returns: %04d, %04d\n",(int)(b->ToggleBreakpoint(40)+1),(int)(b->ToggleBreakpoint(55)+1));
 	//prog8
-	printf("Set Breakpoint returns: %04d, %04d, %04d, %04d, %04d\n",b->ToggleBreakpoint(72)+1,b->ToggleBreakpoint(31)+1,b->ToggleBreakpoint(39)+1,b->ToggleBreakpoint(47)+1,b->ToggleBreakpoint(53)+1);
+	printf("Set Breakpoint returns: %04d, %04d, %04d, %04d, %04d\n",(int)(b->ToggleBreakpoint(72)+1),(int)(b->ToggleBreakpoint(31)+1),(int)(b->ToggleBreakpoint(39)+1),(int)(b->ToggleBreakpoint(47)+1),(int)(b->ToggleBreakpoint(53)+1));
 	system("pause");
 	b->set_DEBUG_ONLY_TESTING(5);
 	while (1) {
